refactor: board단 deprecated 코드 분리
- controller 분리
- service 분리
- dto 분리
