+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/domains/board/controller/BoardController.kt
@@ -10,10 +10,8 @@ import com.yapp.muckpot.common.dto.CursorPaginationRequest
 import com.yapp.muckpot.common.utils.ResponseEntityUtil
 import com.yapp.muckpot.common.utils.SecurityContextHolderUtil
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotCreateRequest
-import com.yapp.muckpot.domains.board.controller.dto.MuckpotCreateRequestV1
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotCreateResponse
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotUpdateRequest
-import com.yapp.muckpot.domains.board.controller.dto.MuckpotUpdateRequestV1
 import com.yapp.muckpot.domains.board.service.BoardService
 import com.yapp.muckpot.domains.user.enums.MuckPotStatus
 import io.swagger.annotations.Api
@@ -226,47 +224,5 @@ class BoardController(
     @GetMapping("/v1/boards/regions")
     fun findAllRegions(): ResponseEntity<ResponseDto> {
         return ResponseEntityUtil.ok(boardService.findAllRegions())
-    }
-
-    @ApiResponses(
-        value = [
-            ApiResponse(
-                code = 201,
-                examples = Example(
-                    ExampleProperty(
-                        value = MUCKPOT_SAVE_RESPONSE,
-                        mediaType = MediaType.APPLICATION_JSON_VALUE
-                    )
-                ),
-                message = "성공"
-            )
-        ]
-    )
-    @ApiOperation(value = "먹팟 글 생성 - v1")
-    @PostMapping("/v1/boards")
-    @Deprecated("V2 배포 후 제거")
-    fun saveBoardV1(
-        @AuthenticationPrincipal userId: Long,
-        @RequestBody @Valid
-        request: MuckpotCreateRequestV1
-    ): ResponseEntity<ResponseDto> {
-        return ResponseEntityUtil.created(
-            MuckpotCreateResponse(
-                boardService.saveBoardV1(userId, request)
-            )
-        )
-    }
-
-    @ApiOperation(value = "먹팟 글 수정 - v1")
-    @PatchMapping("/v1/boards/{boardId}")
-    @Deprecated("V2 배포 후 제거")
-    fun updateBoardV1(
-        @AuthenticationPrincipal userId: Long,
-        @PathVariable boardId: Long,
-        @RequestBody @Valid
-        request: MuckpotUpdateRequestV1
-    ): ResponseEntity<ResponseDto> {
-        boardService.updateBoardAndSendEmailV1(userId, boardId, request)
-        return ResponseEntityUtil.noContent()
     }
 }
