+++ muckpot-api/src/main/kotlin/com/yapp/muckpot/domains/board/service/BoardService.kt
@@ -5,11 +5,9 @@ import com.yapp.muckpot.common.dto.CursorPaginationResponse
 import com.yapp.muckpot.common.redisson.DistributedLock
 import com.yapp.muckpot.domains.board.controller.converter.RegionConverter
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotCreateRequest
-import com.yapp.muckpot.domains.board.controller.dto.MuckpotCreateRequestV1
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotDetailResponse
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotReadResponse
 import com.yapp.muckpot.domains.board.controller.dto.MuckpotUpdateRequest
-import com.yapp.muckpot.domains.board.controller.dto.MuckpotUpdateRequestV1
 import com.yapp.muckpot.domains.board.entity.Participant
 import com.yapp.muckpot.domains.board.exception.BoardErrorCode
 import com.yapp.muckpot.domains.board.exception.ParticipantErrorCode
@@ -196,40 +194,5 @@ class BoardService(
                 muckpotCityResponses.add(RegionConverter.convertToCityResponse(city, provinces))
             }
         return RegionResponse(muckpotCityResponses)
-    }
-
-    @Deprecated("V2 배포 후 제거")
-    @Transactional
-    fun saveBoardV1(userId: Long, request: MuckpotCreateRequestV1): Long? {
-        // TODO 먹팟 등록 시 같은 회사 인원에게 메일 전송
-        val user = userRepository.findByIdOrNull(userId)
-            ?: throw MuckPotException(UserErrorCode.USER_NOT_FOUND)
-        val board = boardRepository.save(request.toBoard(user))
-        participantRepository.save(Participant(user, board))
-        return board.id
-    }
-
-    @Deprecated("V2 배포 후 제거")
-    @Transactional
-    fun updateBoardAndSendEmailV1(userId: Long, boardId: Long, request: MuckpotUpdateRequestV1) {
-        boardRepository.findByIdOrNull(boardId)?.let { board ->
-            if (board.isNotMyBoard(userId)) {
-                throw MuckPotException(BoardErrorCode.BOARD_UNAUTHORIZED)
-            }
-            // Update 이전에 수행되어야 함.
-            val mailTitle = EmailTemplate.BOARD_UPDATE_EMAIL.formatSubject(board.title)
-            val mailBody = EmailTemplate.BOARD_UPDATE_EMAIL.formatBody(
-                board.title,
-                request.createBoardUpdateMailBody(board)
-            )
-            request.updateBoard(board)
-            participantService.sendEmailToParticipantsWithoutWriter(
-                board = board,
-                mailTitle = mailTitle,
-                mailBody = mailBody
-            )
-        } ?: run {
-            throw MuckPotException(BoardErrorCode.BOARD_NOT_FOUND)
-        }
     }
 }
